+++ src/frontend/komponenter/Behandling/Inngangsvilkår/NyttBarnSammePartner/utils.ts
@@ -1,6 +1,5 @@
 import { RegistergrunnlagNyttBarn, SøknadsgrunnlagNyttBarn } from './typer';
-import { IAnnenForelder, IBarnMedSamvær } from '../Aleneomsorg/typer';
-import { formaterNullableFødsesnummer, formaterNullableIsoDato } from '../../../../utils/formatter';
+import { IBarnMedSamvær } from '../Aleneomsorg/typer';
 
 export const mapTilSøknadsgrunnlagNyttBarn = (
     barnMedSamvær: IBarnMedSamvær[]
@@ -31,25 +30,6 @@ export const mapTilRegistergrunnlagNyttBarn = (
                 barn.søknadsgrunnlag.ikkeOppgittAnnenForelderBegrunnelse,
         }))
         .filter((barn) => barn.fødselsnummer);
-};
-
-export const mapForelderTilNavnOgFnr = (forelder?: IAnnenForelder): string | undefined => {
-    if (forelder?.navn === 'ikke oppgitt') return undefined;
-    return [
-        forelder?.navn,
-        forelder?.fødselsnummer
-            ? formaterNullableFødsesnummer(forelder?.fødselsnummer)
-            : formaterNullableIsoDato(forelder?.fødselsdato),
-    ]
-        .filter(Boolean)
-        .join(' - ');
-};
-
-export const mapIkkeOppgitt = (ikkeOppgittAnnenForelderBegrunnelse?: string | null): string => {
-    if (!ikkeOppgittAnnenForelderBegrunnelse) return '';
-    return 'Donor' === ikkeOppgittAnnenForelderBegrunnelse
-        ? 'Donor'
-        : `Kan ikke oppgi: ${ikkeOppgittAnnenForelderBegrunnelse}`;
 };
 
 export const mapBarnNavnTekst = (søknadsgrunnlagNyttBarn: SøknadsgrunnlagNyttBarn): string => {
